+++ lib/src/enums/cron_expression_type.dart
@@ -1,0 +1,1 @@
+enum CronExpressionType { STANDARD, QUARTZ }